+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"24.05.2026 13:33"
-#define NE_VERSION   L"2026.05.24.13"
+#define NE_PUBLISHED L"25.05.2026 12:46"
+#define NE_VERSION   L"2026.05.25.12"